requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.0
PyYAML>=6.0
openai>=1.0.0